refactor(api): isolate MCP account selection helper
Agent-Logs-Url: https://github.com/gredice/gredice/sessions/fdf6a8b0-6822-4638-8b8e-8e844603b5d5
