+++ apps/api/lib/mcp/accountSelection.ts
@@ -1,0 +1,11 @@
+export function resolveMcpAccountId(
+    selectedAccountId: string | undefined,
+    accountIds: string[],
+) {
+    if (selectedAccountId) {
+        return accountIds.includes(selectedAccountId)
+            ? selectedAccountId
+            : null;
+    }
+    return accountIds[0] ?? null;
+}
